Update CopytoTarget method to handle ZipFiles. Add GetBaseDirName method to extract bag names from zipped bags.

+++ src/main/java/nl/knaw/dans/easy/sword2examples/Common.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-17 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
+ * Copyright (C) 2016 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,6 +33,7 @@ import org.apache.abdera.model.Link;
 import org.apache.abdera.parser.Parser;
 import org.apache.commons.codec.binary.Hex;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.HttpEntity;
 import org.apache.http.auth.AuthScope;
@@ -58,7 +59,6 @@ import java.nio.charset.StandardCharsets;
 import java.security.DigestInputStream;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 public class Common {
     static final String BAGIT_URI = "http://purl.org/net/sword/package/BagIt";
@@ -67,11 +67,9 @@ public class Common {
     /**
      * Assumes the entity is UTF-8 encoded text and reads it into a String.
      *
-     * @param entity
-     *        the http entity object
+     * @param entity the http entity object
      * @return the entire http entity as a string
-     * @throws IOException
-     *         if an I/O error occurs
+     * @throws IOException if an I/O error occurs
      */
     public static String readEntityAsString(HttpEntity entity) throws IOException {
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
@@ -123,26 +121,26 @@ public class Common {
                         List<String> dois = getDois(entries.get(0));
                         int numDois = dois.size();
                         switch (numDois) {
-                        case 1:
-                            System.out.print(" With DOI: [" + dois.get(0) + "]. ");
-                            break;
-                        case 0:
-                            System.out.println("WARNING: No DOI found");
-                            break;
+                            case 1:
+                                System.out.print(" With DOI: [" + dois.get(0) + "]. ");
+                                break;
+                            case 0:
+                                System.out.println("WARNING: No DOI found");
+                                break;
 
-                        default:
-                            System.out.println("WARNING: More than one DOI found (" + numDois + "): ");
-                            boolean first = true;
-                            for (String doi : dois) {
-                                if (first)
-                                    first = false;
-                                else
-                                    System.out.print(", ");
-                                System.out.print(doi + "");
+                            default:
+                                System.out.println("WARNING: More than one DOI found (" + numDois + "): ");
+                                boolean first = true;
+                                for (String doi : dois) {
+                                    if (first)
+                                        first = false;
+                                    else
+                                        System.out.print(", ");
+                                    System.out.print(doi + "");
 
-                            }
-                            System.out.println();
-                            break;
+                                }
+                                System.out.println();
+                                break;
                         }
                     } else {
                         System.out.println("WARNING: Found (" + entries.size() + ") entry's; should be ONE and only ONE");
@@ -187,17 +185,16 @@ public class Common {
     }
 
     public static CloseableHttpResponse sendChunk(DigestInputStream dis, int size, String method, URI uri, String filename, String mimeType,
-            CloseableHttpClient http, boolean inProgress) throws Exception
-    {
+                                                  CloseableHttpClient http, boolean inProgress) throws Exception {
         // System.out.println(String.format("Sending chunk to %s, filename = %s, chunk size = %d, MIME-Type = %s, In-Progress = %s ... ", uri.toString(),
         // filename, size, mimeType, Boolean.toString(inProgress)));
         byte[] chunk = readChunk(dis, size);
         String md5 = new String(Hex.encodeHex(dis.getMessageDigest().digest()));
         HttpUriRequest request = RequestBuilder.create(method).setUri(uri).setConfig(RequestConfig.custom()
-            /*
-             * When using an HTTPS-connection EXPECT-CONTINUE must be enabled, otherwise buffer overflow may follow
-             */
-            .setExpectContinueEnabled(true).build()) //
+                /*
+                 * When using an HTTPS-connection EXPECT-CONTINUE must be enabled, otherwise buffer overflow may follow
+                 */
+                .setExpectContinueEnabled(true).build()) //
                 .addHeader("Content-Disposition", String.format("attachment; filename=%s", filename)) //
                 .addHeader("Content-MD5", md5) //
                 .addHeader("Packaging", BAGIT_URI) //
@@ -240,9 +237,25 @@ public class Common {
     }
 
     public static File copyToTarget(File dir) throws Exception {
-        File dirInTarget = new File("target", dir.getName());
-        FileUtils.deleteQuietly(dirInTarget);
-        FileUtils.copyDirectory(dir, dirInTarget);
+        File dirInTarget = null;
+        if (dir.isDirectory()) {
+            dirInTarget = new File("target", dir.getName());
+            FileUtils.deleteQuietly(dirInTarget);
+            FileUtils.copyDirectory(dir, dirInTarget);
+        } else {
+            ZipFile zf = new ZipFile(dir);
+            if (!zf.isValidZipFile()) {
+                System.err.println("The submitted bag is not a valid directory or Zipfile");
+                System.exit(1);
+            } else {
+                File zipInTarget = new File("target", dir.getName());
+                FileUtils.deleteQuietly(zipInTarget);
+                dirInTarget = new File("target", FileUnzipper.getBaseDirName(dir.toString()));
+                FileUtils.deleteQuietly(dirInTarget);
+                zf.extractAll("target");
+            }
+        }
         return dirInTarget;
     }
 }
+
